refactor: extract pallete values into separate file
